파이선 파서 기본 구조 구현
메세지 컨슈머 구현 및 파서 기본 구조 구현
메세지에 따라 파서가 트리거 되도록 함
