+++ batchwos-batch/src/test/java/com/siotman/batchwos/batch/MessageTest.java
@@ -1,0 +1,16 @@
+package com.siotman.batchwos.batch;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@SpringBootTest
+@RunWith(SpringRunner.class)
+public class MessageTest {
+
+
+    @Test
+    public void sendingTest() {
+    }
+}
