--- batchwos-rabbitmq/src/main/java/com/siotman/batchwos/rabbitmq/BatchwosRabbitmqApplication.java
@@ -1,13 +1,0 @@
-package com.siotman.batchwos.rabbitmq;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BatchwosRabbitmqApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BatchwosRabbitmqApplication.class, args);
-    }
-
-}
