--- batchwos-rabbitmq/src/test/java/com/siotman/batchwos/rabbitmq/BatchwosRabbitmqApplicationTests.java
@@ -1,16 +1,0 @@
-package com.siotman.batchwos.rabbitmq;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class BatchwosRabbitmqApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
