--- python-parser/paper_data_dump_0.json
@@ -1,3 +1,0 @@
-{
-    
-}
