+++ python-parser/parser_logger.py
@@ -1,0 +1,6 @@
+class Logger():
+    def __init__(self, name):
+        self.name = name
+    
+    def log(self, level, msg):
+        print("[%s] [%s]"%(level, self.name), msg)
